+++ src/components/QuickSettings.vue
@@ -1,15 +1,18 @@
 <template>
     <div class="row">
         <q-toolbar>
-            <q-chip icon="mdi-brain" :label="isCreateImageSelected ? 'DALL·E' : model" size="sm" clickable
+            <!-- Model name -->
+            <q-chip icon="mdi-brain" :label="isCreateImageSelected ? 'DALL·E 3' : model" size="sm" clickable
                 @click="nextModel()">
                 <q-tooltip :delay="1000" max-width="300px" transition-show="scale" transition-hide="scale">
                     {{ t('settings.openAI.model.label') }}
                 </q-tooltip>
             </q-chip>
+
+            <!-- Max tokens selection -->
             <q-chip v-if="!isCreateImageSelected" :label="maxTokens" size="sm" clickable
-                :icon="maxTokens < 1024 ? 'mdi-text-short' : maxTokens < 2048 ? 'mdi-text' : 'mdi-text-long'" >
-                
+                :icon="maxTokens < 1024 ? 'mdi-text-short' : maxTokens < 2048 ? 'mdi-text' : 'mdi-text-long'">
+
                 <q-menu anchor="top middle" self="bottom middle">
                     <div class="q-pa-sm">
                         <q-slider v-model="maxTokens" :min="64" :max="4096" :step="16" :markers="1024" vertical reverse />
@@ -20,6 +23,7 @@
                 </q-tooltip>
             </q-chip>
 
+            <!-- Text generation temperature -->
             <q-chip v-if="!isCreateImageSelected" icon="mdi-thermometer" :label="temperature" size="sm" clickable>
                 <q-menu anchor="top middle" self="bottom middle">
                     <div class="q-pa-sm">
@@ -33,7 +37,8 @@
 
             <q-separator v-if="!isCreateImageSelected" vertical inset class="q-ma-sm" />
 
-            <q-chip v-if="isCreateImageSelected" icon="mdi-image-multiple-outline" :label="choices" size="sm" clickable>
+            <!-- Image choices selection -->
+            <q-chip v-if="isCreateImageSelected && false" icon="mdi-image-multiple-outline" :label="choices" size="sm" clickable>
                 <q-menu anchor="top middle" self="bottom middle">
                     <div class="q-pa-sm">
                         <q-slider v-model="choices" snap :min="1" :max="10" :step="1" :markers="1" vertical reverse />
@@ -44,6 +49,7 @@
                 </q-tooltip>
             </q-chip>
 
+            <!-- Image size -->
             <q-chip v-if="isCreateImageSelected" icon="mdi-image-size-select-large" :label="imageSize" size="sm" clickable>
                 <q-menu anchor="top middle" self="bottom middle">
                     <div class="q-pa-sm">
@@ -55,6 +61,23 @@
                 </q-tooltip>
             </q-chip>
 
+            <!-- Image quality toggle -->
+            <q-chip v-if="isCreateImageSelected" :icon="imageQuality == 'hd' ? 'mdi-high-definition' : 'mdi-standard-definition'"
+                size="md" clickable @click="imageQuality = imageQuality == 'hd' ? 'standard' : 'hd'" >
+                <q-tooltip :delay="1000" max-width="300px" transition-show="scale" transition-hide="scale">
+                    {{ t('settings.openAI.quality.label') }}
+                </q-tooltip>
+            </q-chip>
+
+            <!-- Image style toggle -->
+            <q-chip v-if="isCreateImageSelected" icon="mdi-palette-outline" :label="imageStyle == 'vivid' ? t('settings.openAI.style.vivid') : t('settings.openAI.style.natural')"
+                size="sm" clickable @click="imageStyle = imageStyle == 'vivid' ? 'natural' : 'vivid'">
+                <q-tooltip :delay="1000" max-width="300px" transition-show="scale" transition-hide="scale">
+                    {{ t('settings.openAI.style.label') }}
+                </q-tooltip>
+            </q-chip>
+
+            <!-- Persona selection -->
             <q-select v-if="appMode == 'advanced' && !isCreateImageSelected" dense options-dense use-chips multiple
                 borderless :option-label="(item) => item === null ? 'Null value' : item.name" v-model="personas"
                 :options="personaOptions" @filter="personaFilterFn">
@@ -96,6 +119,7 @@
 
             <q-separator v-if="appMode == 'advanced' && !isCreateImageSelected" vertical inset class="q-ma-sm" />
 
+            <!-- Team work toggle -->
             <q-chip v-if="appMode == 'advanced' && !isCreateImageSelected" size="sm" icon="mdi-account-group-outline"
                 clickable>
 
@@ -136,6 +160,8 @@ export default {
             temperature,
             choices,
             imageSize,
+            imageQuality,
+            imageStyle,
             personas,
             speechLanguage,
             userAvatar
@@ -158,10 +184,24 @@ export default {
             });
         }
 
-        const imageSizeOptions = ref(['256x256', '512x512', '1024x1024']);
-        const imageSizeValue = ref(imageSizeOptions.value.indexOf(imageSize.value));
+        // Image related
+        const imageSizeOptions = ['1024x1024', '1024x1792', '1792x1024'];
+        const imageQualityOptions = ['standard', 'hd'];
+        const imageStyleOptions = ['vivid', 'natural'];
+        const imageSizeValue = ref(imageSizeOptions.indexOf(imageSize.value));
+        const imageQualityValue = ref(imageQualityOptions.indexOf(imageQuality.value));
+        const imageStyleValue = ref(imageStyleOptions.indexOf(imageStyle.value));
+
         watch(imageSizeValue, () => {
-            imageSize.value = imageSizeOptions.value[imageSizeValue.value];
+            imageSize.value = imageSizeOptions[imageSizeValue.value];
+        });
+
+        watch(imageQualityValue, () => {
+            imageQuality.value = imageQualityOptions[imageQualityValue.value];
+        });
+
+        watch(imageStyleValue, () => {
+            imageStyle.value = imageStyleOptions[imageStyleValue.value];
         });
 
         return {
@@ -176,6 +216,9 @@ export default {
             choices,
             imageSize,
             imageSizeValue,
+            imageQuality,
+            imageQualityValue,
+            imageStyle,
             personas,
             userAvatar,
             isCreateImageSelected,